Add rudimentary querying of items by userId
As a placeholder, I will create an association between the User and the Item model. @ulander-a is currently working on the association of User and MetaUser.

In `lib/Items.js`, I'll create a function that excludes certain columns when querying, in order to shape a response without duplicated or unnecessary data.